Move form picker style separted
