+++ app/components/base/form/create-form/index.ts
@@ -1,4 +1,3 @@
 export * from "./create-form.component";
 export * from "./form-page.component";
-export * from "./form-picker.component";
 export * from "./form-section.component";
