+++ app/components/base/form/form-picker/form-picker.scss
@@ -1,0 +1,59 @@
+@import "app/styles/variables";
+
+bl-form-picker > .form-picker {
+    border: 2px dashed $dusty-grey;
+    min-width: 200px;
+    max-width: 100%;
+    height: 60px;
+    background: transparent;
+    font-size: 16px;
+    color: $dusty-grey;
+    cursor: pointer;
+    border-radius: 5px;
+    outline: none;
+    overflow: hidden;
+    text-overflow: hidden;
+    white-space: nowrap;
+
+    &:hover, &:focus {
+        border-color: map-get($md-prussian-blue, 300);
+        color: map-get($md-prussian-blue, 300);
+    }
+
+    &.has-value {
+        border-style: solid;
+    }
+
+    .fa-plus {
+        font-size: 30px;
+        margin: 0 20px;
+    }
+
+    > .no-value-content {
+        display: flex;
+        align-items: center;
+    }
+
+    > .content {
+        display: flex;
+        align-items: center;
+        > .details {
+            flex: 1;
+            margin: 0 5px;
+            text-overflow: ellipsis;
+            overflow: hidden;
+
+            > * {
+                text-overflow: ellipsis;
+                overflow: hidden;
+            }
+            > .subtitle {
+                font-size: 13px;
+            }
+        }
+
+        > .clear-btn {
+            margin: 0 5px;
+        }
+    }
+}
