+++ app/components/base/form/form-picker/index.ts
@@ -1,0 +1,1 @@
+export * from "./form-picker.component";
